text-to-image-diffuser/requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
datasets>=2.10.0
safetensors
ftfy
einops
pillow
scipy
gradio
omegaconf
pyyaml
tqdm
evaluate
